requirements.txt
easygui~=0.98.2
Pillow~=8.3.2
